LayoutTests/imported/w3c/web-platform-tests/css/css-values/ric-invalidation-expected.txt

CONSOLE MESSAGE: Error: assert_not_equals: expect update of ric units with font size change got disallowed value 784

FAIL CSS Values and Units Test: ric invalidation Error: assert_not_equals: expect update of ric units with font size change got disallowed value 784
